Create day21.py
DAILY CHALLENGE:

Split String - Maximum Value
The program accept a string S containing only Os and 1s as the input.
The program must split the string S into two substring values so that the concatenation of the number of Os in the left substring and the number of 1s in the right substring is maximum. Then the program must print those two substring values of S as the output

DAILY TEST:

Unique Alphabets Sorting

The program must accept a string S containing only lower case alphabets as the input. The program must print the output based on the following conditions.
- The program must print the unique alphabets in the string S in alphabetical order.
- After printing the unique alphabets, the program must remove them from the string S. Then the program must repeat the processes until all the characters in S are removed.
All alphabets must be printed on the same line without space.