Create day22.py
DAILY CHALLENGE:

Remove Pairs - Different in Case
The program must accept a string S as the input. The program must remove two continuous alphabets if they are same but different in case. Similarly, the program must remove the alphabets until there are no two such continuous alphabets in S. If all the characters are removed from the string S, the program must print -1 as the output.

DAILY TEST:

Symmetrical Patterns - Matrix
The program must accept an integer matrix of size RxC as the input. The program must print the output based on the following conditions.
- If the matrix is horizontally symmetric (bottom half of the matrix is equal to water image of the top half), the program must print the string value "Horizontally Symmetric".
- If the matrix is vertically symmetric (right half of the matrix is equal to mirror image of the left half), the program must print the string value "Vertically Symmetric".
- If the matrix is both horizontally and vertically symmetric, the program must print the string value "Both".
- Else the program must print -1 as the output.
Note: If the value of R or C is odd, the middle row or column of the matrix will not be considered when dividing the matrix into two halves (horizontally or vertically).