Create day23.py
DAILY CHALLENGE:

The program must accept a string S and N words as the input. The string S always contains exactly 26 lower case alphabets in any order. The position of an alphabet in the string S represents the score of the alphabet. The program must find the score of each word among the N words by finding the sum of scores of each alphabet in the word. Then the program must print the word which has the maximum score as the output. If two or more words have the same maximum score, the program must print the first occurring word as the output.

DAILY TEST:

The program must accept an integer N as the input. The program must print the next largest integer of N with the same number of 1s as in the binary representation of N.