Create day24.py
DAILY CHALLENGE:

Common Factors (X, Y) and (Y, Z)
The program must accept three integers X, Y and Z as the input. The program must find the common factors of X and Y. Then the program must find the common factors of Y and Z. Finally, the program must print all the factors obtained in descending order as the output.

DAILY TEST:

Valid Mix of Five Alphabets
The program must accept a string S containing the first 5 lower case alphabets (a, b, c, d and e) as the input. The program must print YES if the string S is a valid mix. Else the program must print NO as the output.
If the string S contains multiple "abcde" and the alphabets are mixed in the same order of their occurrence, then it is a valid mix.
Note: The length of S is always a multiple of 5.