Update day25.py
Daily Challenge:

Given an integer N which indicates the position in the alphabet series (upper case) and another integer R which indicates the number of rows to be printed, the program must print R rows with each row containing alphabets equal to the row count.

Daily Test:

The program must accept an integer N as the input. The program must print the pairs of prime integers that differ by 6 in the given range from 2 to N. The pairs must be printed in the order of their occurrence.
Note: Optimize your logic to avoid Time Limit Exceeded error.